+++ CHANGELOG.md
@@ -11,6 +11,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - The public interface into a type accessible via `<instance>.Critr`
 - Term *filter* to *stage* keeping closer to MongoDB's terminology.
+- Operator `$literal` to be a real operator.
+- Operator context property `criteria` changed to `expression`
+- `evaluate` to support general operators.
 
 ### Removed
 - `resetOps` for now. Create new instance of type for a clean state.